* Attempt to fix high ping times by sending response every frame again, but: * Use NETMSG_SYNCRESPONSE for ping calc (since it contains frameNumber anyway)   (This should save about 33% idle (no commands etc.) upload rate of clients   compared to 0.75b2, and 4% compared to 0.76b1.) * Fix float-to-int warning.
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@5244 37977431-3df6-0310-b722-df95706aa16b

+++ rts/Sim/ModInfo.cpp
@@ -53,9 +53,9 @@ void CModInfo::Init(const char* modname)
 
 	// constructions
 	const LuaTable constructionTbl = root.SubTable("construction");
-	constructionDecay       = constructionTbl.GetBool("constructionDecay",  true);
-	constructionDecayTime   = constructionTbl.GetFloat("constructionDecayTime",  6.66) * 30;
-	constructionDecaySpeed  = constructionTbl.GetFloat("constructionDecaySpeed",  0.03);
+	constructionDecay = constructionTbl.GetBool("constructionDecay", true);
+	constructionDecayTime = (int)(constructionTbl.GetFloat("constructionDecayTime", 6.66) * 30);
+	constructionDecaySpeed = constructionTbl.GetFloat("constructionDecaySpeed", 0.03);
 
 	// reclaim
 	const LuaTable reclaimTbl = root.SubTable("reclaim");
